ApadtiveLearningAssessments
This is a basic version of a program that tracks students accuracy is certain topics and reenforces topics that they struggle with.

In detail:
Throws random questions at the student.
Based on results, for the next quiz a specified percent of the questions are for topics the student has a low accuracy rate at getting correct while the remainder is random to make sure the student still understands the rest of the material.

Added features could be:
     Only look at recent answers from the student so that if they get a bunch wrong in the beginning, that does not affect their progress later on.
     Also implement the difficulty ratings based on students answers. Starting off easy and then progressing to harder questions as they answer correctly.