Made changes to auto populate link field
The Quiz Component will now look for a link argument in the URL which will then auto-populate the link field on the dashboard.

Was able to get the button working while hosting the Video2AI site publicly on my computer with ngrok.

Also am adding a vercel.json (used for allowing most recent version of react) file to try and setup hosting with vercel so I do not need to host it publicly on our personal machine.