Dynamic Login/Louout Button
Button now dynamically shows login or logout depending on the clerk user login state.

Also got rid of new error with :redirectUrl as a prop when passing down to the Button.

Another error is known when logging out: Only plain objects, and a few built-ins, can be passed to Client Components from Server Components. Classes or null prototypes are not supported.
  {}