Chat Session Bug Fix
Fixed the wording of the chat session for requesting the quiz questions since sometimes the format was off.

Also now when there is a failure generating the quiz questions, they are not in the correct format or any other error in fetching them, (usually due to bad internet) it should retry 5 times before quitting and allowing the user to retry again if they choose to. Also prints an error message to the page to let the user know.

Also if there is no connection at all to be able to connect to Gemini, usually due to Wifi being turned off completely, the same quiz they just took will be presented.

Also some small wording changed in ChangeTranscript Page for better tutor directions.